swc_ecma_regexp_visit

Trait Visit

pub trait Visit {

    // Provided methods
    fn visit_alternative(&mut self, node: &Alternative) { ... }
    fn visit_alternatives(&mut self, node: &[Alternative]) { ... }
    fn visit_atom(&mut self, node: &Atom) { ... }
    fn visit_boundary_assertion(&mut self, node: &BoundaryAssertion) { ... }
    fn visit_boundary_assertion_kind(&mut self, node: &BoundaryAssertionKind) { ... }
    fn visit_capturing_group(&mut self, node: &CapturingGroup) { ... }
    fn visit_character(&mut self, node: &Character) { ... }
    fn visit_character_class(&mut self, node: &CharacterClass) { ... }
    fn visit_character_class_contents(&mut self, node: &CharacterClassContents) { ... }
    fn visit_character_class_contents_kind(
        &mut self,
        node: &CharacterClassContentsKind,
    ) { ... }
    fn visit_character_class_contentss(
        &mut self,
        node: &[CharacterClassContents],
    ) { ... }
    fn visit_character_class_escape(&mut self, node: &CharacterClassEscape) { ... }
    fn visit_character_class_escape_kind(
        &mut self,
        node: &CharacterClassEscapeKind,
    ) { ... }
    fn visit_character_class_range(&mut self, node: &CharacterClassRange) { ... }
    fn visit_character_kind(&mut self, node: &CharacterKind) { ... }
    fn visit_characters(&mut self, node: &[Character]) { ... }
    fn visit_class_string(&mut self, node: &ClassString) { ... }
    fn visit_class_string_disjunction(&mut self, node: &ClassStringDisjunction) { ... }
    fn visit_class_strings(&mut self, node: &[ClassString]) { ... }
    fn visit_disjunction(&mut self, node: &Disjunction) { ... }
    fn visit_dot(&mut self, node: &Dot) { ... }
    fn visit_ignore_group(&mut self, node: &IgnoreGroup) { ... }
    fn visit_indexed_reference(&mut self, node: &IndexedReference) { ... }
    fn visit_look_around_assertion(&mut self, node: &LookAroundAssertion) { ... }
    fn visit_look_around_assertion_kind(
        &mut self,
        node: &LookAroundAssertionKind,
    ) { ... }
    fn visit_modifier(&mut self, node: &Modifier) { ... }
    fn visit_modifiers(&mut self, node: &Modifiers) { ... }
    fn visit_named_reference(&mut self, node: &NamedReference) { ... }
    fn visit_opt_atom(&mut self, node: &Option<Atom>) { ... }
    fn visit_opt_modifiers(&mut self, node: &Option<Modifiers>) { ... }
    fn visit_pattern(&mut self, node: &Pattern) { ... }
    fn visit_quantifier(&mut self, node: &Quantifier) { ... }
    fn visit_span(&mut self, node: &Span) { ... }
    fn visit_term(&mut self, node: &Term) { ... }
    fn visit_terms(&mut self, node: &[Term]) { ... }
    fn visit_unicode_property_escape(&mut self, node: &UnicodePropertyEscape) { ... }

}

A visitor trait for traversing the AST.

Provided Methods

fn visit_alternative(&mut self, node: &Alternative)

Visit a node of type Alternative.

By default, this method calls Alternative::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_alternatives(&mut self, node: &[Alternative])

Visit a node of type Vec < Alternative >.

By default, this method calls [Vec < Alternative >::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_atom(&mut self, node: &Atom)

Visit a node of type swc_atoms :: Atom.

By default, this method calls [swc_atoms :: Atom::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_boundary_assertion(&mut self, node: &BoundaryAssertion)

Visit a node of type BoundaryAssertion.

By default, this method calls BoundaryAssertion::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_boundary_assertion_kind(&mut self, node: &BoundaryAssertionKind)

Visit a node of type BoundaryAssertionKind.

By default, this method calls BoundaryAssertionKind::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_capturing_group(&mut self, node: &CapturingGroup)

Visit a node of type CapturingGroup.

By default, this method calls CapturingGroup::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character(&mut self, node: &Character)

Visit a node of type Character.

By default, this method calls Character::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character_class(&mut self, node: &CharacterClass)

Visit a node of type CharacterClass.

By default, this method calls CharacterClass::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character_class_contents(&mut self, node: &CharacterClassContents)

Visit a node of type CharacterClassContents.

By default, this method calls CharacterClassContents::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character_class_contents_kind( &mut self, node: &CharacterClassContentsKind, )

Visit a node of type CharacterClassContentsKind.

By default, this method calls CharacterClassContentsKind::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character_class_contentss(&mut self, node: &[CharacterClassContents])

Visit a node of type Vec < CharacterClassContents >.

By default, this method calls [Vec < CharacterClassContents >::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_character_class_escape(&mut self, node: &CharacterClassEscape)

Visit a node of type CharacterClassEscape.

By default, this method calls CharacterClassEscape::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character_class_escape_kind(&mut self, node: &CharacterClassEscapeKind)

Visit a node of type CharacterClassEscapeKind.

By default, this method calls CharacterClassEscapeKind::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character_class_range(&mut self, node: &CharacterClassRange)

Visit a node of type CharacterClassRange.

By default, this method calls CharacterClassRange::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_character_kind(&mut self, node: &CharacterKind)

Visit a node of type CharacterKind.

By default, this method calls CharacterKind::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_characters(&mut self, node: &[Character])

Visit a node of type Vec < Character >.

By default, this method calls [Vec < Character >::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_class_string(&mut self, node: &ClassString)

Visit a node of type ClassString.

By default, this method calls ClassString::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_class_string_disjunction(&mut self, node: &ClassStringDisjunction)

Visit a node of type ClassStringDisjunction.

By default, this method calls ClassStringDisjunction::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_class_strings(&mut self, node: &[ClassString])

Visit a node of type Vec < ClassString >.

By default, this method calls [Vec < ClassString >::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_disjunction(&mut self, node: &Disjunction)

Visit a node of type Disjunction.

By default, this method calls Disjunction::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_dot(&mut self, node: &Dot)

Visit a node of type Dot.

By default, this method calls Dot::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_ignore_group(&mut self, node: &IgnoreGroup)

Visit a node of type IgnoreGroup.

By default, this method calls IgnoreGroup::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_indexed_reference(&mut self, node: &IndexedReference)

Visit a node of type IndexedReference.

By default, this method calls IndexedReference::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_look_around_assertion(&mut self, node: &LookAroundAssertion)

Visit a node of type LookAroundAssertion.

By default, this method calls LookAroundAssertion::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_look_around_assertion_kind(&mut self, node: &LookAroundAssertionKind)

Visit a node of type LookAroundAssertionKind.

By default, this method calls LookAroundAssertionKind::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_modifier(&mut self, node: &Modifier)

Visit a node of type Modifier.

By default, this method calls Modifier::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_modifiers(&mut self, node: &Modifiers)

Visit a node of type Modifiers.

By default, this method calls Modifiers::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_named_reference(&mut self, node: &NamedReference)

Visit a node of type NamedReference.

By default, this method calls NamedReference::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_opt_atom(&mut self, node: &Option<Atom>)

Visit a node of type Option < swc_atoms :: Atom >.

By default, this method calls [Option < swc_atoms :: Atom >::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_opt_modifiers(&mut self, node: &Option<Modifiers>)

Visit a node of type Option < Modifiers >.

By default, this method calls [Option < Modifiers >::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_pattern(&mut self, node: &Pattern)

Visit a node of type Pattern.

By default, this method calls Pattern::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_quantifier(&mut self, node: &Quantifier)

Visit a node of type Quantifier.

By default, this method calls Quantifier::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_span(&mut self, node: &Span)

Visit a node of type swc_common :: Span.

By default, this method calls [swc_common :: Span::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_term(&mut self, node: &Term)

Visit a node of type Term.

By default, this method calls Term::visit_children_with. If you want to recurse, you need to call it manually.

fn visit_terms(&mut self, node: &[Term])

Visit a node of type Vec < Term >.

By default, this method calls [Vec < Term >::visit_children_with]. If you want to recurse, you need to call it manually.

fn visit_unicode_property_escape(&mut self, node: &UnicodePropertyEscape)

Visit a node of type UnicodePropertyEscape.

By default, this method calls UnicodePropertyEscape::visit_children_with. If you want to recurse, you need to call it manually.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".