swc_ecma_regexp_visit

Trait VisitAstPath

pub trait VisitAstPath {

    // Provided methods
    fn visit_alternative<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Alternative,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_alternatives<'ast: 'r, 'r>(
        &mut self,
        node: &'ast [Alternative],
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_atom<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Atom,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_boundary_assertion<'ast: 'r, 'r>(
        &mut self,
        node: &'ast BoundaryAssertion,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_boundary_assertion_kind<'ast: 'r, 'r>(
        &mut self,
        node: &'ast BoundaryAssertionKind,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_capturing_group<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CapturingGroup,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Character,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_class<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CharacterClass,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_class_contents<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CharacterClassContents,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_class_contents_kind<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CharacterClassContentsKind,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_class_contentss<'ast: 'r, 'r>(
        &mut self,
        node: &'ast [CharacterClassContents],
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_class_escape<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CharacterClassEscape,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_class_escape_kind<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CharacterClassEscapeKind,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_class_range<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CharacterClassRange,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_character_kind<'ast: 'r, 'r>(
        &mut self,
        node: &'ast CharacterKind,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_characters<'ast: 'r, 'r>(
        &mut self,
        node: &'ast [Character],
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_class_string<'ast: 'r, 'r>(
        &mut self,
        node: &'ast ClassString,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_class_string_disjunction<'ast: 'r, 'r>(
        &mut self,
        node: &'ast ClassStringDisjunction,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_class_strings<'ast: 'r, 'r>(
        &mut self,
        node: &'ast [ClassString],
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_disjunction<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Disjunction,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_dot<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Dot,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_ignore_group<'ast: 'r, 'r>(
        &mut self,
        node: &'ast IgnoreGroup,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_indexed_reference<'ast: 'r, 'r>(
        &mut self,
        node: &'ast IndexedReference,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_look_around_assertion<'ast: 'r, 'r>(
        &mut self,
        node: &'ast LookAroundAssertion,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_look_around_assertion_kind<'ast: 'r, 'r>(
        &mut self,
        node: &'ast LookAroundAssertionKind,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_modifier<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Modifier,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_modifiers<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Modifiers,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_named_reference<'ast: 'r, 'r>(
        &mut self,
        node: &'ast NamedReference,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_opt_atom<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Option<Atom>,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_opt_modifiers<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Option<Modifiers>,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_pattern<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Pattern,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_quantifier<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Quantifier,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_span<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Span,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_term<'ast: 'r, 'r>(
        &mut self,
        node: &'ast Term,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_terms<'ast: 'r, 'r>(
        &mut self,
        node: &'ast [Term],
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }
    fn visit_unicode_property_escape<'ast: 'r, 'r>(
        &mut self,
        node: &'ast UnicodePropertyEscape,
        __ast_path: &mut AstNodePath<'r>,
    ) { ... }

}

Available on crate feature path only.

A visitor trait for traversing the AST.

Provided Methods

fn visit_alternative<'ast: 'r, 'r>( &mut self, node: &'ast Alternative, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Alternative.

By default, this method calls [Alternative::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_alternatives<'ast: 'r, 'r>( &mut self, node: &'ast [Alternative], __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Vec < Alternative >.

By default, this method calls [Vec < Alternative >::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_atom<'ast: 'r, 'r>( &mut self, node: &'ast Atom, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type swc_atoms :: Atom.

By default, this method calls [swc_atoms :: Atom::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_boundary_assertion<'ast: 'r, 'r>( &mut self, node: &'ast BoundaryAssertion, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type BoundaryAssertion.

By default, this method calls [BoundaryAssertion::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_boundary_assertion_kind<'ast: 'r, 'r>( &mut self, node: &'ast BoundaryAssertionKind, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type BoundaryAssertionKind.

By default, this method calls [BoundaryAssertionKind::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_capturing_group<'ast: 'r, 'r>( &mut self, node: &'ast CapturingGroup, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CapturingGroup.

By default, this method calls [CapturingGroup::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character<'ast: 'r, 'r>( &mut self, node: &'ast Character, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Character.

By default, this method calls [Character::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_class<'ast: 'r, 'r>( &mut self, node: &'ast CharacterClass, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CharacterClass.

By default, this method calls [CharacterClass::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_class_contents<'ast: 'r, 'r>( &mut self, node: &'ast CharacterClassContents, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CharacterClassContents.

By default, this method calls [CharacterClassContents::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_class_contents_kind<'ast: 'r, 'r>( &mut self, node: &'ast CharacterClassContentsKind, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CharacterClassContentsKind.

By default, this method calls [CharacterClassContentsKind::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_class_contentss<'ast: 'r, 'r>( &mut self, node: &'ast [CharacterClassContents], __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Vec < CharacterClassContents >.

By default, this method calls [Vec < CharacterClassContents >::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_class_escape<'ast: 'r, 'r>( &mut self, node: &'ast CharacterClassEscape, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CharacterClassEscape.

By default, this method calls [CharacterClassEscape::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_class_escape_kind<'ast: 'r, 'r>( &mut self, node: &'ast CharacterClassEscapeKind, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CharacterClassEscapeKind.

By default, this method calls [CharacterClassEscapeKind::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_class_range<'ast: 'r, 'r>( &mut self, node: &'ast CharacterClassRange, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CharacterClassRange.

By default, this method calls [CharacterClassRange::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_character_kind<'ast: 'r, 'r>( &mut self, node: &'ast CharacterKind, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type CharacterKind.

By default, this method calls [CharacterKind::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_characters<'ast: 'r, 'r>( &mut self, node: &'ast [Character], __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Vec < Character >.

By default, this method calls [Vec < Character >::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_class_string<'ast: 'r, 'r>( &mut self, node: &'ast ClassString, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type ClassString.

By default, this method calls [ClassString::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_class_string_disjunction<'ast: 'r, 'r>( &mut self, node: &'ast ClassStringDisjunction, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type ClassStringDisjunction.

By default, this method calls [ClassStringDisjunction::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_class_strings<'ast: 'r, 'r>( &mut self, node: &'ast [ClassString], __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Vec < ClassString >.

By default, this method calls [Vec < ClassString >::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_disjunction<'ast: 'r, 'r>( &mut self, node: &'ast Disjunction, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Disjunction.

By default, this method calls [Disjunction::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_dot<'ast: 'r, 'r>( &mut self, node: &'ast Dot, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Dot.

By default, this method calls [Dot::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_ignore_group<'ast: 'r, 'r>( &mut self, node: &'ast IgnoreGroup, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type IgnoreGroup.

By default, this method calls [IgnoreGroup::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_indexed_reference<'ast: 'r, 'r>( &mut self, node: &'ast IndexedReference, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type IndexedReference.

By default, this method calls [IndexedReference::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_look_around_assertion<'ast: 'r, 'r>( &mut self, node: &'ast LookAroundAssertion, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type LookAroundAssertion.

By default, this method calls [LookAroundAssertion::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_look_around_assertion_kind<'ast: 'r, 'r>( &mut self, node: &'ast LookAroundAssertionKind, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type LookAroundAssertionKind.

By default, this method calls [LookAroundAssertionKind::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_modifier<'ast: 'r, 'r>( &mut self, node: &'ast Modifier, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Modifier.

By default, this method calls [Modifier::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_modifiers<'ast: 'r, 'r>( &mut self, node: &'ast Modifiers, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Modifiers.

By default, this method calls [Modifiers::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_named_reference<'ast: 'r, 'r>( &mut self, node: &'ast NamedReference, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type NamedReference.

By default, this method calls [NamedReference::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_opt_atom<'ast: 'r, 'r>( &mut self, node: &'ast Option<Atom>, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Option < swc_atoms :: Atom >.

By default, this method calls [Option < swc_atoms :: Atom >::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_opt_modifiers<'ast: 'r, 'r>( &mut self, node: &'ast Option<Modifiers>, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Option < Modifiers >.

By default, this method calls [Option < Modifiers >::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_pattern<'ast: 'r, 'r>( &mut self, node: &'ast Pattern, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Pattern.

By default, this method calls [Pattern::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_quantifier<'ast: 'r, 'r>( &mut self, node: &'ast Quantifier, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Quantifier.

By default, this method calls [Quantifier::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_span<'ast: 'r, 'r>( &mut self, node: &'ast Span, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type swc_common :: Span.

By default, this method calls [swc_common :: Span::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_term<'ast: 'r, 'r>( &mut self, node: &'ast Term, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Term.

By default, this method calls [Term::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_terms<'ast: 'r, 'r>( &mut self, node: &'ast [Term], __ast_path: &mut AstNodePath<'r>, )

Visit a node of type Vec < Term >.

By default, this method calls [Vec < Term >::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.

fn visit_unicode_property_escape<'ast: 'r, 'r>( &mut self, node: &'ast UnicodePropertyEscape, __ast_path: &mut AstNodePath<'r>, )

Visit a node of type UnicodePropertyEscape.

By default, this method calls [UnicodePropertyEscape::visit_children_with_ast_path]. If you want to recurse, you need to call it manually.