swc_ecma_regexp_visit

Trait VisitMut

pub trait VisitMut {

    // Provided methods
    fn visit_mut_alternative(&mut self, node: &mut Alternative) { ... }
    fn visit_mut_alternatives(&mut self, node: &mut Vec<Alternative>) { ... }
    fn visit_mut_atom(&mut self, node: &mut Atom) { ... }
    fn visit_mut_boundary_assertion(&mut self, node: &mut BoundaryAssertion) { ... }
    fn visit_mut_boundary_assertion_kind(
        &mut self,
        node: &mut BoundaryAssertionKind,
    ) { ... }
    fn visit_mut_capturing_group(&mut self, node: &mut CapturingGroup) { ... }
    fn visit_mut_character(&mut self, node: &mut Character) { ... }
    fn visit_mut_character_class(&mut self, node: &mut CharacterClass) { ... }
    fn visit_mut_character_class_contents(
        &mut self,
        node: &mut CharacterClassContents,
    ) { ... }
    fn visit_mut_character_class_contents_kind(
        &mut self,
        node: &mut CharacterClassContentsKind,
    ) { ... }
    fn visit_mut_character_class_contentss(
        &mut self,
        node: &mut Vec<CharacterClassContents>,
    ) { ... }
    fn visit_mut_character_class_escape(
        &mut self,
        node: &mut CharacterClassEscape,
    ) { ... }
    fn visit_mut_character_class_escape_kind(
        &mut self,
        node: &mut CharacterClassEscapeKind,
    ) { ... }
    fn visit_mut_character_class_range(
        &mut self,
        node: &mut CharacterClassRange,
    ) { ... }
    fn visit_mut_character_kind(&mut self, node: &mut CharacterKind) { ... }
    fn visit_mut_characters(&mut self, node: &mut Vec<Character>) { ... }
    fn visit_mut_class_string(&mut self, node: &mut ClassString) { ... }
    fn visit_mut_class_string_disjunction(
        &mut self,
        node: &mut ClassStringDisjunction,
    ) { ... }
    fn visit_mut_class_strings(&mut self, node: &mut Vec<ClassString>) { ... }
    fn visit_mut_disjunction(&mut self, node: &mut Disjunction) { ... }
    fn visit_mut_dot(&mut self, node: &mut Dot) { ... }
    fn visit_mut_ignore_group(&mut self, node: &mut IgnoreGroup) { ... }
    fn visit_mut_indexed_reference(&mut self, node: &mut IndexedReference) { ... }
    fn visit_mut_look_around_assertion(
        &mut self,
        node: &mut LookAroundAssertion,
    ) { ... }
    fn visit_mut_look_around_assertion_kind(
        &mut self,
        node: &mut LookAroundAssertionKind,
    ) { ... }
    fn visit_mut_modifier(&mut self, node: &mut Modifier) { ... }
    fn visit_mut_modifiers(&mut self, node: &mut Modifiers) { ... }
    fn visit_mut_named_reference(&mut self, node: &mut NamedReference) { ... }
    fn visit_mut_opt_atom(&mut self, node: &mut Option<Atom>) { ... }
    fn visit_mut_opt_modifiers(&mut self, node: &mut Option<Modifiers>) { ... }
    fn visit_mut_pattern(&mut self, node: &mut Pattern) { ... }
    fn visit_mut_quantifier(&mut self, node: &mut Quantifier) { ... }
    fn visit_mut_span(&mut self, node: &mut Span) { ... }
    fn visit_mut_term(&mut self, node: &mut Term) { ... }
    fn visit_mut_terms(&mut self, node: &mut Vec<Term>) { ... }
    fn visit_mut_unicode_property_escape(
        &mut self,
        node: &mut UnicodePropertyEscape,
    ) { ... }

}

A visitor trait for traversing the AST.

Provided Methods

fn visit_mut_alternative(&mut self, node: &mut Alternative)

Visit a node of type Alternative.

By default, this method calls Alternative::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_alternatives(&mut self, node: &mut Vec<Alternative>)

Visit a node of type Vec < Alternative >.

By default, this method calls [Vec < Alternative >::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_atom(&mut self, node: &mut Atom)

Visit a node of type swc_atoms :: Atom.

By default, this method calls [swc_atoms :: Atom::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_boundary_assertion(&mut self, node: &mut BoundaryAssertion)

Visit a node of type BoundaryAssertion.

By default, this method calls BoundaryAssertion::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_boundary_assertion_kind( &mut self, node: &mut BoundaryAssertionKind, )

Visit a node of type BoundaryAssertionKind.

By default, this method calls BoundaryAssertionKind::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_capturing_group(&mut self, node: &mut CapturingGroup)

Visit a node of type CapturingGroup.

By default, this method calls CapturingGroup::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character(&mut self, node: &mut Character)

Visit a node of type Character.

By default, this method calls Character::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character_class(&mut self, node: &mut CharacterClass)

Visit a node of type CharacterClass.

By default, this method calls CharacterClass::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character_class_contents( &mut self, node: &mut CharacterClassContents, )

Visit a node of type CharacterClassContents.

By default, this method calls CharacterClassContents::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character_class_contents_kind( &mut self, node: &mut CharacterClassContentsKind, )

Visit a node of type CharacterClassContentsKind.

By default, this method calls CharacterClassContentsKind::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character_class_contentss( &mut self, node: &mut Vec<CharacterClassContents>, )

Visit a node of type Vec < CharacterClassContents >.

By default, this method calls [Vec < CharacterClassContents >::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_character_class_escape(&mut self, node: &mut CharacterClassEscape)

Visit a node of type CharacterClassEscape.

By default, this method calls CharacterClassEscape::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character_class_escape_kind( &mut self, node: &mut CharacterClassEscapeKind, )

Visit a node of type CharacterClassEscapeKind.

By default, this method calls CharacterClassEscapeKind::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character_class_range(&mut self, node: &mut CharacterClassRange)

Visit a node of type CharacterClassRange.

By default, this method calls CharacterClassRange::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_character_kind(&mut self, node: &mut CharacterKind)

Visit a node of type CharacterKind.

By default, this method calls CharacterKind::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_characters(&mut self, node: &mut Vec<Character>)

Visit a node of type Vec < Character >.

By default, this method calls [Vec < Character >::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_class_string(&mut self, node: &mut ClassString)

Visit a node of type ClassString.

By default, this method calls ClassString::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_class_string_disjunction( &mut self, node: &mut ClassStringDisjunction, )

Visit a node of type ClassStringDisjunction.

By default, this method calls ClassStringDisjunction::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_class_strings(&mut self, node: &mut Vec<ClassString>)

Visit a node of type Vec < ClassString >.

By default, this method calls [Vec < ClassString >::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_disjunction(&mut self, node: &mut Disjunction)

Visit a node of type Disjunction.

By default, this method calls Disjunction::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_dot(&mut self, node: &mut Dot)

Visit a node of type Dot.

By default, this method calls Dot::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_ignore_group(&mut self, node: &mut IgnoreGroup)

Visit a node of type IgnoreGroup.

By default, this method calls IgnoreGroup::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_indexed_reference(&mut self, node: &mut IndexedReference)

Visit a node of type IndexedReference.

By default, this method calls IndexedReference::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_look_around_assertion(&mut self, node: &mut LookAroundAssertion)

Visit a node of type LookAroundAssertion.

By default, this method calls LookAroundAssertion::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_look_around_assertion_kind( &mut self, node: &mut LookAroundAssertionKind, )

Visit a node of type LookAroundAssertionKind.

By default, this method calls LookAroundAssertionKind::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_modifier(&mut self, node: &mut Modifier)

Visit a node of type Modifier.

By default, this method calls Modifier::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_modifiers(&mut self, node: &mut Modifiers)

Visit a node of type Modifiers.

By default, this method calls Modifiers::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_named_reference(&mut self, node: &mut NamedReference)

Visit a node of type NamedReference.

By default, this method calls NamedReference::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_opt_atom(&mut self, node: &mut Option<Atom>)

Visit a node of type Option < swc_atoms :: Atom >.

By default, this method calls [Option < swc_atoms :: Atom >::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_opt_modifiers(&mut self, node: &mut Option<Modifiers>)

Visit a node of type Option < Modifiers >.

By default, this method calls [Option < Modifiers >::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_pattern(&mut self, node: &mut Pattern)

Visit a node of type Pattern.

By default, this method calls Pattern::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_quantifier(&mut self, node: &mut Quantifier)

Visit a node of type Quantifier.

By default, this method calls Quantifier::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_span(&mut self, node: &mut Span)

Visit a node of type swc_common :: Span.

By default, this method calls [swc_common :: Span::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_term(&mut self, node: &mut Term)

Visit a node of type Term.

By default, this method calls Term::visit_mut_children_with. If you want to recurse, you need to call it manually.

fn visit_mut_terms(&mut self, node: &mut Vec<Term>)

Visit a node of type Vec < Term >.

By default, this method calls [Vec < Term >::visit_mut_children_with]. If you want to recurse, you need to call it manually.

fn visit_mut_unicode_property_escape( &mut self, node: &mut UnicodePropertyEscape, )

Visit a node of type UnicodePropertyEscape.

By default, this method calls UnicodePropertyEscape::visit_mut_children_with. If you want to recurse, you need to call it manually.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".